requirements.txt
Django==5.1.1
django-crispy-forms==2.3
crispy-bootstrap5==2024.2
django_debug_toolbar==4.4.6
django-environ==0.11.2
django_extensions==3.2.3
django_simple_plausible==0.0.4
django-singleton==0.1.9
gunicorn==23.0.0
ipython==8.27.0
Pillow==10.4.0
